Added edge tracking + debug display for the Background_Controller

+++ Tunnelour/Tunnelour/include/Avatar_Controller.h
@@ -161,7 +161,7 @@ class Avatar_Controller: public Tunnelour::Controller {
   //             : Tileset_Helper::Collision_Block
   //---------------------------------------------------------------------------
   Avatar_Component::Collision_Block TilesetCollisionBlockToAvatarCollisionBlock(Tileset_Helper::Collision_Block tileset_collision_block,
-                                                                                int tileset_animation_top_left_y,
+                                                                                float tileset_animation_top_left_y,
                                                                                 int state_index);
 
   //---------------------------------------------------------------------------
